pyproject.toml
[tool.poetry]
name = "pix-r-picks-backend"
version = "0.1.0"
description = ""
authors = ["Kassie Bradshaw <kassie.r.bradshaw@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.9"
django-environ = "^0.8.1"
djangorestframework = "^3.12.4"
text2emotion = "^0.0.5"
jupyterlab = "^3.2.1"
PyPDF2 = "^1.26.0"
psycopg2-binary = "^2.9.1"
djangorestframework-simplejwt = "^5.0.0"
gunicorn = "^20.1.0"
whitenoise = "^5.3.0"
django-cors-headers = "^3.10.0"
testcase = "^0.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
